backend/requirements.txt
azure-core
azure-identity
python-dotenv
pyaudio
azure-ai-voicelive
SpeechRecognition
